+++ changelog.md
@@ -4,7 +4,8 @@ For the latest version of this document, please see [https://github.com/haskell/
 
 - Change time instances of types with year (`Day`, `UTCTime`) to require years with at least 4 digits.
 - Change `KeyValue` instances to be more general (and use equality to constraint them) instead of being more lax flexible instances.
-- Export `Key` type also from `Data.Aeson.KeyMap` module
+- Export `Key` type also from `Data.Aeson.KeyMap` module.
+- Export `mapWithKey` from `Data.Aeson.KeyMap` module.
 
 ### 2.0.3.0